Handle validation in resolvers

+++ server/mutations/hall/deleteHall.js
@@ -1,4 +1,6 @@
 const Hall = require('../../models/hall')
-module.exports = async (root, { id }) => {
-    const hall = await Hall.findByIdAndDelete(id);
+module.exports = async (root, { id }, { user }) => {
+    if (!user.isAdmin) throw new Error('You are not allowed to do access this resource !!')
+
+    return await Hall.findByIdAndDelete(id);
 }
