+++ server/mutations/module/deleteModule.js
@@ -1,4 +1,5 @@
 const Module = require('../../models/module')
-module.exports = async (root, { id }) => {
-    const module = await Module.findOneAndDelete({ moduleId: id });
+module.exports = async (root, { id }, { user }) => {
+    if (!user.isAdmin) throw new Error('You are not allowed to do access this resource !!')
+    return await Module.findOneAndDelete({ moduleId: id });
 }
