+++ server/mutations/student_module/deleteStudentModule.js
@@ -1,5 +1,10 @@
 const Student_Module = require('../../models/student_module')
-module.exports = async (root, { input }) => {
+module.exports = async (root, { input }, { user }) => {
+
+    if (!user.isAdmin) throw new Error('You are not allowed to do access this resource !!')
+
     const { studentId, moduleId, isEligible } = input
-    const student_module = await Student_Module.findOneAndDelete({ studentId: studentId, moduleId: moduleId, isEligible: isEligible });
+    
+    return await Student_Module.findOneAndDelete({ studentId: studentId, moduleId: moduleId, isEligible: isEligible });
+
 }
