+++ server/mutations/user/deleteUser.js
@@ -1,5 +1,9 @@
 const User = require('../../models/user')
-module.exports = async (root, { input }) => {
+module.exports = async (root, { input }, {user}) => {
+    
+    if (!user.isAdmin) throw new Error('You are not allowed to do access this resource !!')
+
     const { id } = input;
-    const user = await User.findOneAndDelete({ studentId: id });
+    
+    return await User.findOneAndDelete({ studentId: id });
 }
